requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-geoposition==0.3.0
django-google-maps==0.13.0
django-health-check==3.18.3
django-location-field==2.7.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
